third_party/blink/web_tests/svg/dynamic-updates/SVGFEComponentTransferElement-svgdom-intercept-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'intercept' property of the SVGFEComponentTransferElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Math.round(feRFunc.intercept.baseVal * 100) is 20
PASS Math.round(feGFunc.intercept.baseVal * 100) is 20
PASS Math.round(feBFunc.intercept.baseVal * 100) is 20
PASS Math.round(feAFunc.intercept.baseVal * 100) is 20
PASS Math.round(feRFunc.intercept.baseVal * 100) is 10
PASS Math.round(feGFunc.intercept.baseVal * 100) is 10
PASS Math.round(feBFunc.intercept.baseVal * 100) is 10
PASS Math.round(feAFunc.intercept.baseVal * 100) is 10
PASS successfullyParsed is true

TEST COMPLETE
